refactor(errors): error index module
