+++ src/dictionaries/ShaKeyAlgorithm.js
@@ -4,6 +4,7 @@
 const crypto = require('crypto')
 const KeyAlgorithm = require('./KeyAlgorithm')
 const {TextEncoder, TextDecoder} = require('text-encoding')
+const {OperationError} = require('../errors')
 
 /**
  * ShaKeyAlgorithm
