+++ src/errors/index.js
@@ -1,0 +1,9 @@
+module.exports = {
+  DataError: require('./DataError'),
+  InvalidAccessError: require('./InvalidAccessError'),
+  KeyFormatNotSupportedError: require('./KeyFormatNotSupportedError'),
+  NotSupportedError: require('./NotSupportedError'),
+  OperationError: require('./OperationError'),
+  QuotaExceededError: require('./QuotaExceededError'),
+  TypeMismatchError: require('./TypeMismatchError')
+}
